src/libs/algolia/fetchAlgolia/buildVenuesQueryOptions.ts
import { AlgoliaLocationFilter, FiltersArray } from 'libs/algolia'
import { VenuesFacets } from 'libs/algolia/enums'
import { getVenueTypeFacetFilters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/getVenueTypeFacetFilters'

type AlgoliaVenueQueryOptions = {
  tags?: string[]
  venueTypes?: string[]
  locationFilter: AlgoliaLocationFilter
}

export const buildVenuesQueryOptions = ({
  tags = [],
  venueTypes = [],
  locationFilter,
}: AlgoliaVenueQueryOptions) => {
  const facetFilters: FiltersArray = []

  if (tags.length) {
    const tagsPredicate = buildTagsPredicate(tags)
    facetFilters.push(tagsPredicate)
  }

  if (venueTypes.length) {
    const venueTypesPredicate = buildVenueTypesPredicate(venueTypes.map(getVenueTypeFacetFilters))
    facetFilters.push(venueTypesPredicate)
  }

  // We want to show on home page only venues that have at least one offer that is searchable in algolia
  const hasAtLeastOneBookableOfferPredicate = [
    `${VenuesFacets.has_at_least_one_bookable_offer}:true`,
  ]
  facetFilters.push(hasAtLeastOneBookableOfferPredicate)

  return {
    locationFilter,
    ...(facetFilters.length > 0 ? { facetFilters } : {}),
  }
}

const buildVenueTypesPredicate = (venueTypes: string[]): string[] =>
  venueTypes.map((venueType) => `${VenuesFacets.venue_type}:${venueType}`)

const buildTagsPredicate = (tags: string[]): string[] =>
  tags.map((tag: string) => `${VenuesFacets.tags}:${tag}`)
